Block placement update
Added block placement speed. Added block placement condition. Removed the ability to create a block inside an entity. Removed the ability to create a block inside a tree. Remove the ability to create a block inside a pick up item. Remove the ability to mine a block under a tree.